extract TabsViewHolder from Tabs model
